+++ src/index.ts
@@ -1,7 +1,7 @@
-/// <reference types="typescript" />
-"use strict";
+import {AsyncState as _AsyncState}      from "./AsyncState";
 
-import AsyncState from "./AsyncState";
+export {AsyncState}      from "./AsyncState";
+export {AsyncStateGroup} from "./AsyncStateGroup";
 
 export * from "./Types";
 export * from "./Utils";
@@ -11,4 +11,4 @@ export * from "./Updater";
 export * from "./react";
 export * from "./redux";
 
-export default AsyncState;
+export default _AsyncState;